% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{allocate_cards}
\alias{allocate_cards}
\title{fonction qui repartie les cartes pour les afficher sur la table}
\usage{
allocate_cards(cards)
}
\arguments{
\item{une}{liste d'idcards}
}
\value{
une liste de 3 vecteurs ( pour les 3 lignes affichees)
}
\description{
fonction qui repartie les cartes pour les afficher sur la table
}

