% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_and.R
\name{and2}
\alias{and2}
\alias{\%and\%}
\alias{\%a\%}
\alias{\%A\%}
\alias{&}
\title{Get Intersection Set for Two Sets}
\usage{
a \%and\% b

a \%a\% b

a \%A\% b

"&"(a, b)
}
\arguments{
\item{a}{see argument x in \code{\link[set]{toVector}}}

\item{b}{see argument x in \code{\link[set]{toVector}}}
}
\value{
intersection set
}
\description{
Get intersection set for two sets, which can be numbers, characters, vectors even dataframe, matrix or list.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
A \%and\% B
A \%and\% B \%and\% C
}
