% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/atomizeSets.R
\name{atomizeSets}
\alias{atomizeSets}
\title{A function to make a mosaic-style plot for sets and their atoms.}
\usage{
atomizeSets(sets, values = NULL, col = "red", bgCol = "ivory3",
  orderBy = "none", setLab = NULL, atomAvLab = NULL, ylab = "Atom size")
}
\arguments{
\item{sets}{List representing sets of features.}

\item{values}{Optional non-negative value (or weight) assigned to each feature. In the case that it is NULL, all features have the same weight.}

\item{col}{Color representing sets and their respective atoms.}

\item{bgCol}{Background color.}

\item{orderBy}{If equal to "size", atoms are ordered by their sizes; if equal to "mean", atoms are ordered by the mean values assigned to their features.}

\item{setLab}{Vector of set labels; if NULL, use the names of the sets list.}

\item{atomAvLab}{Character representing labels above atoms in case the "mean" option is used in orderBy above; if NULL, use "Average in atom".}

\item{ylab}{Character representing y-axis label.}
}
\description{
This function allows you to input a list of sets of features and visualize their atoms via a mosaic-style plot.
}
\examples{
sets <- list(A=1:50, B=31:130, C=131:160, D=c(40:50, 161:260))
atomizeSets(sets)
}

