\name{ChromSets}
\docType{data}
\alias{sets}

\title{Gene-sets from Subramanian et al. study}
\description{List of gene-sets used in Subramanian et al. study,
  representing chromosomal regions. 40 of the original 212 were
  excluded, in order to obtain non-overlapping sets (atoms.)}

\usage{data(ChromSets)}
\format{
  List of gene-sets, with the names representing the set names
  and the entries representing the vector of genes in each set.
}

\references{
Subramanian, A., P. Tamayo, V. Mootha, S. Mukherjee, B. Ebert, M. Gillette,
A. Paulovich, S. Pomeroy, T. Golub, E. Lander, et al. (2005): Gene set enrichment
analysis: A knowledge-based approach for interpreting genome-wide expression
profiles, \emph{Proceedings of the National Academy of Sciences}, 102, 15545-5550.
}

\keyword{datasets}
