% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{combination}
\alias{combination}
\title{Combination of Characters or Vectors}
\usage{
combination(...)
}
\arguments{
\item{...}{one or more vectors}
}
\value{
binary combination
}
\description{
Combination of characters or vectors.
}
\examples{
A <- c("a","b","c")
combination(A)

B <- c("a","b","c","d")
C <- c("a","e","h")
D <- c("a","b","e")
E <- c("a","c")
combination(A,B)
combination(A,B,C)
combination(A,B,C,D)
combination(A,B,C,D,E)
}
