% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{draw_card}
\alias{draw_card}
\title{dibujar una carta}
\usage{
draw_card(forme, nombre, remplissage, couleur)
}
\arguments{
\item{forme}{(1 : rectangle, 2: circle, 3:losange)}

\item{nombre}{(1/2/3)}

\item{remplissage}{(vide / plein / hachure)}

\item{couleur}{(1/2/3)}
}
\value{
nothing
}
\description{
dibujar una carta
}

