% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{draw_cards}
\alias{draw_cards}
\title{draw ncards randomly}
\usage{
draw_cards(ncards, jeu)
}
\arguments{
\item{ncards}{: an integer and jeu : a data.frame with the game play}
}
\value{
a list of 2 data.frame : drawn_cards and remaining_cards
}
\description{
draw ncards randomly
}
\examples{
draw_cards( 12, create_gameplay() )

}

