% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{draw_game}
\alias{draw_game}
\title{create un juego de SET! 81 cards in .png}
\usage{
draw_game(jeu)
}
\arguments{
\item{jeu}{data.frame created by create_game}
}
\value{
time elapsed
}
\description{
create un juego de SET! 81 cards in .png
}

