% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{egal_}
\alias{egal_}
\title{fonction teste si un critere est egal pour les 3 cartes et renvoie TRUE ou FALSE}
\usage{
egal_(trio)
}
\arguments{
\item{trio}{un dataframe de 3 lignes et 1 colonne}
}
\value{
TRUE if all values are equals FALSE otherwise
}
\description{
fonction teste si un critere est egal pour les 3 cartes et renvoie TRUE ou FALSE
}

