\name{GenderStats}
\docType{data}
\alias{gene.stat}

\title{Gene-level statistics for data from Subramanian et al. study}
\description{
  Gene-level statistics for a dataset comparing mRNA expression
  profiles from 15 males and 17 females from lymphoblastoid cell lines
  from the Subramanian et al. study. We note that the data is
  approximately normal (Irizarry et al.)
}

\usage{data(GenderStats)}
\format{
  Vector of gene-level statistics.
}

\references{
Subramanian, A., P. Tamayo, V. Mootha, S. Mukherjee, B. Ebert, M. Gillette,
A. Paulovich, S. Pomeroy, T. Golub, E. Lander, et al. (2005): Gene set enrichment
analysis: A knowledge-based approach for interpreting genome-wide expression
profiles, \emph{Proceedings of the National Academy of Sciences}, 102,
15545-15550.

Irizarry, R., C. Wang, Y. Zhou, and T. Speed (2009):
Gene set enrichment analysis
made simple, \emph{JHU Biostat Working Paper 185},
\url{http://www.bepress.com/jhubiostat/paper185/}
}

\keyword{datasets}
