% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_comp.R
\name{\%comp\%}
\alias{\%comp\%}
\title{Get Complementary Elements}
\usage{
a \%comp\% b
}
\arguments{
\item{a}{must be a subset of b, which can be numbers, characters, vectors, one dataframe, one matrix or one list.}

\item{b}{can be numbers, characters, vectors, one dataframe, one matrix or one list.}
}
\value{
complementary elements in data set b
}
\description{
Get complementary elements between data set a and b.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
A \%comp\% B
}
