% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_not.R
\name{\%not\%}
\alias{\%not\%}
\title{Get Elements only Existed in Dataset a}
\usage{
a \%not\% b
}
\arguments{
\item{a}{can be numbers, characters, vectors, one dataframe, one matrix or one list.}

\item{b}{can be numbers, characters, vectors, one dataframe, one matrix or one list.}
}
\value{
elements only existed in dataset a
}
\description{
Get elements only existed in dataset a.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
B \%not\% A
}
