% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_or.R
\name{\%or\%}
\alias{\%or\%}
\title{Get Union Set for Two Sets}
\usage{
a \%or\% b
}
\arguments{
\item{a}{can be numbers, characters, vectors, one dataframe, one matrix or one list.}

\item{b}{can be numbers, characters, vectors, one dataframe, one matrix or one list.}
}
\value{
union set
}
\description{
Get union set for two sets.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
A \%or\% B
A \%and\% B \%or\% C
}
