% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_sub.R
\name{\%sub\%}
\alias{\%sub\%}
\title{Subset}
\usage{
a \%sub\% A
}
\arguments{
\item{a}{subset, which can be numbers, characters, vectors, one dataframe, one matrix or one list.}

\item{A}{data set, which can be numbers, characters, vectors, one dataframe, one matrix or one list.}
}
\value{
logical result.
}
\description{
Whether data set a is a subject of data set A.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
A \%sub\% B
}
