% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.R
\name{grep}
\alias{grep}
\alias{grep_or}
\alias{grep_and}
\alias{grep_not_and}
\alias{grep_not_or}
\title{Grep for vectors}
\usage{
grep_or(x, patterns)

grep_and(x, patterns)

grep_not_and(x, patterns)

grep_not_or(x, patterns)
}
\arguments{
\item{x}{vectors with one or more characters}

\item{patterns}{vectors with one or more characters}
}
\value{
vectors
}
\description{
Grep for vectors
}
\examples{
x=c('a12','a','b')
patterns=c('b','1')
grep_or(x,patterns)
x=c('a12','a1','b')
patterns=c('a','1')
grep_and(x,patterns)
x=c('a12','a','b')
patterns=c('a','1')
grep_not_and(x,patterns)
x=c('a12','a','b')
patterns=c('a','1')
grep_not_or(x,patterns)
}
