% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_not_and_or.R
\name{grepl}
\alias{grepl}
\alias{grepl_or}
\alias{grepl_and}
\alias{grepl_not_and}
\alias{grepl_not_or}
\title{Grepl for vectors}
\usage{
grepl_or(x, patterns)

grepl_and(x, patterns)

grepl_not_and(x, patterns)

grepl_not_or(x, patterns)
}
\arguments{
\item{x}{vectors with one or more characters}

\item{patterns}{vectors with one or more characters}
}
\value{
vectors
}
\description{
Grepl for vectors
}
\examples{
x=c('a12','a','b')
patterns=c('b','1')
grepl_or(x,patterns)
x=c('a12','a1','b')
patterns=c('a','1')
grepl_and(x,patterns)
x=c('a12','a','b')
patterns=c('a','1')
grepl_not_and(x,patterns)
x=c('a12','a','b')
patterns=c('a','1')
grepl_not_or(x,patterns)
}
