% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sub.R
\name{is.sub}
\alias{is.sub}
\title{Judge Subset}
\usage{
is.sub(a, A)
}
\arguments{
\item{a}{see argument x in \code{\link[set]{toVector}}}

\item{A}{see argument x in \code{\link[set]{toVector}}}
}
\value{
logical result.
}
\description{
Whether data set a is a subject of data set A.
}
\examples{
a <- c("a","b","c")
A <- c("a","b","c","d")
is.sub(a,A)
}
