% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.R
\name{not}
\alias{not}
\alias{\%not\%}
\title{Get Elements only Existed in Dataset a}
\usage{
not(...)

a \%not\% b
}
\arguments{
\item{...}{one or more objects}

\item{a}{one object}

\item{b}{the other object}
}
\value{
elements only existed in dataset a
}
\description{
Get elements only existed in dataset a.
}
\examples{
A <- c("a","b")
B <- c("a","b","c","d")
not(B, A)

E <- c('d')
not(B, A, E)
A <- c("a","b","c")
B <- c("a","b","c","d")
B \%not\% A
}
