% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_not.R
\name{not2}
\alias{not2}
\alias{\%not\%}
\alias{\%n\%}
\alias{\%N\%}
\alias{/}
\title{Get Elements only Existed in Dataset a}
\usage{
a \%not\% b

a \%n\% b

a \%N\% b

"/"(a, b)
}
\arguments{
\item{a}{see argument x in \code{\link[set]{toVector}}}

\item{b}{see argument x in \code{\link[set]{toVector}}}
}
\value{
elements only existed in dataset a
}
\description{
Get elements only existed in dataset a.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
B \%not\% A
}
