% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepes_or.R
\name{or2}
\alias{or2}
\alias{\%or\%}
\alias{\%r\%}
\alias{\%R\%}
\alias{|}
\title{Get Union Set for Two Sets}
\usage{
a \%or\% b

a \%r\% b

a \%R\% b

"|"(a, b)
}
\arguments{
\item{a}{see argument x in \code{\link[set]{toVector}}}

\item{b}{see argument x in \code{\link[set]{toVector}}}
}
\value{
union set
}
\description{
Get union set for two sets.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
A \%or\% B
A \%and\% B \%or\% C
}
