% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toVector.R
\name{toVector}
\alias{toVector}
\title{Convet to Character}
\usage{
toVector(x, as.numeric = FALSE)
}
\arguments{
\item{x}{can be vector, dataframe, matrix, list, array}

\item{as.numeric}{logical, wheter to numeric form}
}
\value{
a character or numeric vector
}
\description{
Convert dataframe, matrix, list, array or vector to character vector.
}
\examples{
df=data.frame(a=c(1,2,3))
toVector(df)
}
