% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_and.R
\name{and}
\alias{and}
\title{Get Intersection Set for Sets}
\usage{
and(...)
}
\arguments{
\item{...}{see argument x in \code{\link[set]{toVector}}}
}
\value{
intersection elements
}
\description{
Get intersection set for sets.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
and(A, B)
and(A, B, C)
}
