% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{draw_card}
\alias{draw_card}
\title{create un juego de SET! 81 cards in .png}
\usage{
draw_card(forme, nombre, remplissage, couleur)
}
\arguments{
\item{forme}{(1 : rectangle, 2: circle, 3:losange), nombre (1/2/3), remplissage (vide / plein / hachure), couleur (1/2/3)}
}
\value{
nothing
}
\description{
create un juego de SET! 81 cards in .png
}

