% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{or}
\alias{or}
\alias{\%or\%}
\title{Get Union Set for Sets}
\usage{
or(...)

a \%or\% b
}
\arguments{
\item{...}{one or more objects}

\item{a}{one object}

\item{b}{the other object}
}
\value{
union elements
}
\description{
Get union set for sets.
}
\examples{
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
or(A, B)
or(A, B, C)
A <- c("a","b","c")
B <- c("a","b","c","d")
C <- c("a","e","h")
A \%or\% B
A \%and\% B \%or\% C
}
